Adding StageManager
Have always needed similar utilities in lots of projects - decided to
write one that would be bulletproof and extremely robust so there's
little need to expand it out later. Still needs a lot of doc and
testing work before it's ready for primetime but it should be good to
start developing against.